+++ runner/src/server/plugins/engine/configureEnginePlugin.ts
@@ -6,7 +6,7 @@ import { idFromFilename } from "./helpers";
 import {
   loadPreConfiguredForms,
   FormConfiguration,
-} from "./services/configuration-service";
+} from "./services/configurationService";
 
 type ConfigureEnginePlugin = (
   configFile?: string,
